Bug admin (#22)
* admin cannot delete their own accounts and admin panel does not show up for non admins when making a new entry

* changed size of profile picture when viewing account and added error handlers. Also removed console.error and console.logs

* removed missed console.log

* fixed validation bugs

* fixed logger bugs and added a function that creates the log file if it does not already exist.

* added helmet protection

* added safe inline scripting with nonces

* removed unused features

+++ controllers/userController.js
@@ -277,20 +277,17 @@ exports.getEditUser = async function (req, res, next) {
 		res.render("edit-user", user)
 	}
 	catch(error) {
-		console.log("Could not retrieve entries: ", error); 
-		res.redirect("/")
+		next(error); 
 	}
 }
 
-exports.getUser = async function (req, res) {
+exports.getUser = async function (req, res, next) {
 	try {
         const userId = req.query.id;
         const [user] = await userModel.getAccountEntry(userId);
-        console.log([user]);
         res.render("view-user", user);
     } catch (error) {
-        console.log("Could not retrieve user: ", error);
-        res.redirect("/");
+		next(error);
     }
 }
 
@@ -315,7 +312,6 @@ exports.confirmEditUser = async function(req, res, next) {
         email: req.body.email,
         phone_number: req.body.phone_number
     }
-
 
 	const userId = req.body.id; 
 	const adminUserId = req.session.userId;
@@ -344,8 +340,15 @@ exports.confirmEditUser = async function(req, res, next) {
 }
 
 exports.deleteUser = async function (req, res, next) {
-	var userId = req.query.id;
+	const userId = req.query.id;
+	const adminUserId = req.session.userId; 
+
 	try {
+
+		if (userId === adminUserId) {
+			throw new Error('Unauthorized access');
+		}
+
 		await userModel.deleteUser(userId);
 
 		logger.log({
@@ -362,6 +365,6 @@ exports.deleteUser = async function (req, res, next) {
 		res.redirect('/admin-panel'); 
 	}
 	catch (error) {
-    next(error) 
+    	next(error) 
 	}
 }
